+++ esphome/components/display/display_color_utils.h
